eatgo-login-api/src/test/java/com/lyg/eatgo/application/UserServiceTest.java
package com.lyg.eatgo.application;

import com.lyg.eatgo.domain.User;
import com.lyg.eatgo.domain.UserRepository;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.springframework.security.crypto.password.PasswordEncoder;

import java.util.Optional;

import static org.assertj.core.api.Assertions.assertThatThrownBy;
import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.core.Is.is;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.BDDMockito.given;
import static org.mockito.Mockito.never;
import static org.mockito.Mockito.verify;

class UserServiceTest {

    private UserService userService;

    @Mock
    private UserRepository userRepository;

    @Mock
    private PasswordEncoder passwordEncoder;



    @BeforeEach
    public void setUp() {
        MockitoAnnotations.initMocks(this);

        userService = new UserService(userRepository,passwordEncoder);
    }



    @Test
    public void authenticateWithNotExistedEmail() {
        String email = "x@example.com";
        String password = "test";

        assertThatThrownBy(()-> {
            given(userRepository.findByEmail(email)).willReturn(Optional.empty());

            userService.authenticate(email,password);

        }).isInstanceOf(EmailNotExistedException.class);

    }


    @Test
    public void authenticateWithValidAttributes() {
        String email = "tester@example.com";
        String password = "test";

        User mockUser = User.builder().email(email).build();

        given(userRepository.findByEmail(email))
                .willReturn(Optional.of(mockUser));

        given(passwordEncoder.matches(any(), any())).willReturn(true);

        User user = userService.authenticate(email, password);

        assertThat(user.getEmail(), is(email));

    }

    @Test
    public void authenticateWithWrongPassword() {
        String email = "tester@example.com";
        String password = "x";

        User mockUser = User.builder().email(email).build();

        assertThatThrownBy(()-> {
            given(userRepository.findByEmail(email)).willReturn(Optional.of(mockUser));
            userService.authenticate(email,password);

            given(passwordEncoder.matches(any(), any())).willReturn(false);

        }).isInstanceOf(PasswordWrongException.class);
    }
}